requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-filter>=23.0
django-cors-headers>=4.0
pandas>=2.0
python-dotenv>=1.0
whitenoise>=6.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
requests>=2.31.0
